Refactor NavigationHelper to FragmentNavigator
- Relocated `NavigationHelper.kt` from `utils/` to `ui/components/`.
- Renamed `NavigationHelper.kt` to `FragmentNavigator.kt` and updated the object name.
- Updated the package declaration to `org.ole.planet.myplanet.ui.components`.
- Replaced all usages and imports of `NavigationHelper` with `FragmentNavigator` across the codebase.
- Verified build succeeds.
