+++ app/src/main/java/org/ole/planet/myplanet/base/BaseExamFragment.kt
@@ -38,7 +38,7 @@ import org.ole.planet.myplanet.ui.exam.UserInformationFragment
 import org.ole.planet.myplanet.ui.surveys.SurveyFragment
 import org.ole.planet.myplanet.utils.CameraUtils
 import org.ole.planet.myplanet.utils.CameraUtils.ImageCaptureCallback
-import org.ole.planet.myplanet.ui.components.NavigationHelper
+import org.ole.planet.myplanet.utils.NavigationHelper
 import org.ole.planet.myplanet.utils.NetworkUtils.getUniqueIdentifier
 import org.ole.planet.myplanet.utils.Utilities
