+++ app/src/main/java/org/ole/planet/myplanet/ui/community/CommunityServicesFragment.kt
@@ -15,7 +15,7 @@ import org.ole.planet.myplanet.model.RealmNews
 import org.ole.planet.myplanet.ui.teams.TeamDetailFragment
 import org.ole.planet.myplanet.utils.MarkdownUtils.prependBaseUrlToImages
 import org.ole.planet.myplanet.utils.MarkdownUtils.setMarkdownText
-import org.ole.planet.myplanet.ui.components.NavigationHelper.replaceFragment
+import org.ole.planet.myplanet.utils.NavigationHelper.replaceFragment
 
 class CommunityServicesFragment : BaseTeamFragment() {
     private var binding: FragmentCommunityServicesBinding? = null
