+++ app/src/main/java/org/ole/planet/myplanet/ui/teams/members/MembersAdapter.kt
@@ -22,7 +22,7 @@ import org.ole.planet.myplanet.callback.OnMemberActionListener
 import org.ole.planet.myplanet.databinding.RowJoinedUserBinding
 import org.ole.planet.myplanet.repository.JoinedMemberData
 import org.ole.planet.myplanet.utils.DiffUtils
-import org.ole.planet.myplanet.ui.components.NavigationHelper
+import org.ole.planet.myplanet.utils.NavigationHelper
 
 class MembersAdapter(
     private val context: Context,
